Maps/Warnings: reduce the amount of warnings/err...

+++ src/tools/ToolsCata/map_extractor/System.cpp
@@ -305,7 +305,7 @@ uint32 ReadMapDBC()
 
     SFileCloseFile(dbcFile);
     printf("Done! (%u maps loaded)\n", uint32(map_count));
-    return map_count;
+    return static_cast<uint32_t>(map_count);
 }
 
 void ReadAreaTableDBC()
